Move get_plugin_config() to extras.plugins.utils
