+++ netbox/extras/plugins/utils.py
@@ -1,8 +1,10 @@
 from django.apps import apps
 from django.conf import settings
+from django.core.exceptions import ImproperlyConfigured
 
 __all__ = (
     'get_installed_plugins',
+    'get_plugin_config',
 )
 
 
@@ -17,3 +19,19 @@ def get_installed_plugins():
         plugins[plugin_name] = getattr(plugin_config, 'version', None)
 
     return dict(sorted(plugins.items()))
+
+
+def get_plugin_config(plugin_name, parameter, default=None):
+    """
+    Return the value of the specified plugin configuration parameter.
+
+    Args:
+        plugin_name: The name of the plugin
+        parameter: The name of the configuration parameter
+        default: The value to return if the parameter is not defined (default: None)
+    """
+    try:
+        plugin_config = settings.PLUGINS_CONFIG[plugin_name]
+        return plugin_config.get(parameter, default)
+    except KeyError:
+        raise ImproperlyConfigured(f"Plugin {plugin_name} is not registered.")
